Species service
Added SpeciesService

+++ fanbase-web-api/src/main/java/pl/edu/pjwstk/fanbasewebapi/modelDTO/SpeciesDTO.java
@@ -13,7 +13,7 @@ public class SpeciesDTO {
     private double averageHeight;
     private long averageLifespan;
     private String language;
-    private int originPlanetId;
+    private Long originPlanetId;
     private Long swqpiId;
 
 }
